refactor: move fp conversion test to pytest
Moves the floating point conversion tests from the combined_test.py script to a new pytest-style test file.

The new file is located at python/test/unit/language/test_fp_conversion.py.

The run_tests.ps1 script has been updated to run the new test.
