+++ run_tests.ps1
@@ -3,4 +3,4 @@
 [System.Environment]::SetEnvironmentVariable('Path', 'C:\Windows\System32;C:\Program Files\Python312;C:\Program Files\Python312\Scripts;C:\Program Files (x86)\Microsoft Visual Studio\2022\BuildTools\Common7\IDE\CommonExtensions\Microsoft\CMake\CMake\bin;C:\Program Files (x86)\Microsoft Visual Studio\2022\BuildTools\VC\Tools\MSVC\14.44.35207\bin\Hostx64\x64;C:\Program Files (x86)\Windows Kits\10\bin\10.0.20348.0\x64;C:\Program Files\Git\cmd', 'Process')
 [System.Environment]::SetEnvironmentVariable('INCLUDE', 'C:\Program Files (x86)\Microsoft Visual Studio\2022\BuildTools\VC\Tools\MSVC\14.44.35207\include;C:\Program Files (x86)\Windows Kits\10\Include\10.0.20348.0\shared;C:\Program Files (x86)\Windows Kits\10\Include\10.0.20348.0\ucrt;C:\Program Files (x86)\Windows Kits\10\Include\10.0.20348.0\um;C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v12.9\include;C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v12.9\extras\CUPTI\include', 'Process')
 [System.Environment]::SetEnvironmentVariable('LIB', 'C:\Program Files (x86)\Microsoft Visual Studio\2022\BuildTools\VC\Tools\MSVC\14.44.35207\lib\x64;C:\Program Files (x86)\Windows Kits\10\Lib\10.0.20348.0\ucrt\x64;C:\Program Files (x86)\Windows Kits\10\Lib\10.0.20348.0\um\x64', 'Process')
-& "C:\Program Files\Python312\python.exe" python/test/unit/combined_test.py *>&1 | Out-File -FilePath "test_output.log" -Encoding utf8
+& "C:\Program Files\Python312\python.exe" -m pytest python/test/unit/language/test_fp_conversion.py *>&1 | Out-File -FilePath "test_output.log" -Encoding utf8
